+++ src/filters/DraftFilter.php
@@ -10,13 +10,21 @@ use craft\helpers\ElementHelper;
  * Filters events based on whether the element is a draft
  *
  * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
- * @since 2.1
+ * @since 2.1.0
  */
 class DraftFilter extends BaseElementFilter
 {
     public static function displayName(): string
     {
         return Craft::t('webhooks', 'Element is a draft');
+    }
+
+    public static function excludes(): array
+    {
+        return [
+            RevisionFilter::class,
+            FirstSaveFilter::class,
+        ];
     }
 
     protected static function checkElement(ElementInterface $element, bool $value): bool