+++ src/migrations/m200617_212615_debounce_keys.php
@@ -2,7 +2,6 @@
 
 namespace craft\webhooks\migrations;
 
-use Craft;
 use craft\db\Migration;
 
 /**
@@ -13,17 +12,18 @@ class m200617_212615_debounce_keys extends Migration
     /**
      * @inheritdoc
      */
-    public function safeUp()
+    public function safeUp(): bool
     {
         $this->addColumn('{{%webhooks}}', 'debounceKeyFormat', $this->string()->after('filters'));
         $this->addColumn('{{%webhookrequests}}', 'debounceKey', $this->string()->after('webhookId'));
         $this->createIndex(null, '{{%webhookrequests}}', ['debounceKey', 'status']);
+        return true;
     }
 
     /**
      * @inheritdoc
      */
-    public function safeDown()
+    public function safeDown(): bool
     {
         echo "m200617_212615_debounce_keys cannot be reverted.\n";
         return false;